Switch over to use the MDP framework from zeroMQ examples.
This will allow us to have heartbeat workers with services that
they register for. So that way we can tell 3D meshers from 2D meshers

Lots of work needs to be done to add a clean api to the worker code
so that we can extend it for a general use case.