Added the ability to termiante jobs, and redesigned comunication.
The communicaton between the client and server has been changed
to use more objects, rather than the flags we used before. This
should make if far clearer on how you request jobs, and how
you get the status of jobs that are currently running.

We also fixed a couple bugs in the server code when dealing with
numerous jobs pending for workers while the factory has dispatched
multiple workers. The factory created workers would mark jobs
as taken when creatd, and also when connecting, causing some jobs
to never be executed. Also fixed a bug in the Worker pool that
we would remove the wrong worker.

I have also added comments to the majority of the code base, making
it clearer on what and how everything is used.

Change-Id: I5fdd0dca60c13061aa9236b927c1ff5e976b9b89