Update the code to support zmq 2 and 3.
zmq 3 move zmq.hpp into its own project. So we provide a copy of the
bindings ourselves instead of asking people to get another external
project. We also extended the c++ bindings to handle interfacing with
zmq 2 and zmq 3.

Lastly the polling interval has changed in zmq 3, moving from
microseconds to milliseconds which requires some wrapping int the bindings.
The poll command now expects milliseconds, and will properly convert
that to microseconds when dealing with zmq2.