SignalCatcher is a new class that allows classes to catch C signals.
To properly handle abnormal termination of servers/clients/etc we need
a nice way to hook into C signals and relay them to classes that contain
the information required to cleanup everything.

In this case we are going to extend remus::server::Server to inherit
from SignalCatcher to make sure we can properly kill any zombie workers.