Refactor the Job classes to be in client and worker directory.
Rather than each of the Job classes have methods that worked only
on the client or worker, we move them to being unique classes that
reside in the respective folder/namespace.

This allows us to easily extend each side of the api, and allows
a more targeted approach to make the client and worker api.

Also this should allow us in the future to create a better wire protocol
format.

Change-Id: I2ce21b8fa9b1ad4f0f1c8fc7191cf25ec6fa49dc