JobQueue stores JobRequests instead of Messages.
The original prupose of the JobQueue was not only to hold a job, but to keep
it in the zmq layer as long as possible to reduce memory savings.

But because it remus a worker job class it is already removing the data
from zmq and than having it pass it back to zmq.

If we forgo trying to keep data in zmq as long as possible and accept
that remus isn't going to be super super high frequency we should adjust
our api to accept JobRequests.

Change-Id: I46098edc3e1c2ffae4849f4679f8b420e03155a3