Remus worker now support an async set of calls to get a job from the server.
Previously the getJob method was the only way to get a job from
the server, and it was a blocking call. Now you can submit a request
for the number of jobs you want, you can query how many jobs you have pending
and you can fetch the jobs now.

Change-Id: I49bfdbaefb98873a118d16a99c40d20c2bd1328e