WorkerPool now dispatches jobs to multiple workers of the same type better.
Before we also gave a job to the first worker that said it could handle
it, which preferred workers that joined first. Now once a worker takes
a job it goes to the back of the line. This will allow us to load balance
jobs better between workers that all want multiple jobs.

Change-Id: I824859c55800d65d09ecc5788c21828c721dd61a