Fix Visual Studio type narrowing warnings.
Visual Studio is far more aggressive about type narrowing than other compilers.
It should be noted that the biggest change of this is that we have
found that zmq_poll takes a system specific long type which means that
our max polling duration on windows is significantly smaller than on other
operating systems. That means on windows we should move to zmq 3 asap.

Change-Id: Iddef276e993aa2045304272502d661e88f9a7aaf