Allow a user to change the range at which a server can poll at.
When trying to create highly responsive servers that need to shutdown fast,
we need the ability to set a really low poll interval range so that the
poll operation doesn't block the destruction for a long time.

This will also allow users to create servers that run in immediate mode,
if they just have to burn cpu cycles.

Change-Id: Ib7a76b66c0ab1f4cd742c17e1022de42d86ab5dc