MessageRouters zmq::socket are now thread scoped.
Basically we need to the socket lifespan to be the same as the thread
that is doing the polling so that zmq doesn't improperly call recv internally
on a thread will it is being cleaned up.

More information on why this is done can be found at:
  http://zeromq.org/whitepapers:0mq-termination#toc3

This also fixes the issue report in #128.