Workers Test now uses a non standard tcp-ip port.
Previously we had issues with tests happening in parallel and the Worker Test
trying to bind to a tcp-ip port that was being used by another test, or was
in the process of being released by a test. We fix the issue by using a port
that won't be normally used by any other test.