Worker's MessageRouter is now conservative on when it expects to hb again.
Instead of stating a worker will heartbeat with expected regularity, we now
state the next hb will come in as late as possible. The reason for this
is that if we are sending or receiving a massive message ( 5MB+ ) we stall
the hb thread and cause the server to think we are expired.