SubmitJobs tests now reduce generated binary data memory footprint
We want to reduce the time to run the SubmitJobs integration tests, so we cache
the single binary blob to send to the workers. At the same time we keep the
amount of memory the client uses as low as possible to reduce the chance of
OOM issue on travis testing machines.