remus::Worker now blocks while sending results back to the server.
When sending information back to the server, it was possible to delete the
worker before the results had left the workers zmq inbox. This caused the result
message to never reach the server, making the server think the job was still
ongoing.

The criteria for this to happen was that both the server had to be under heavy
load, and the job results be of very large size so that transmission wasn't
instantaneous.

The solution was to add result tracking to the MessageRouter, and require
the server to send a response to the worker when it has received the job result.