Implemented a faster name conversion for the socket Identity.
Instead of having to convert the binary string over to a md5 sum, we convert it based on
if zmq created the name or if the user did so.

If zmq created the name we know that it is a big endiain encoded unsigned integer, so we
can quickly cast it back to that. If the user created the name we presume that it is a textual
unique identitfier.