Fixes Issue #234. Workers heart-beat rate is now customizable.
The primary motivation for this is that for some application workers require
a high level of responsiveness, and a timeout for a dead worker of 1 minute
would make programs look like they had crashed, when it was just a dead worker.

Now people can create very high, or very very slow heart-beat rate workers on a
per instance basis, and have no issues when being connected to an arbitrary
remus server.